Issue #SC-11888 Feat: Scripts to change the primary key in device_profile
