+++ druid/adhoc-jobs/src/main/resources/DeviceProfile.conf
@@ -1,0 +1,4 @@
+spark.cassandra.connection.host="__cassandra__host"
+cassandra.keyspace="__env__"
+cassandra.deviceprofileOldTable="device_profile"
+cassandra.deviceprofileNewTable="device_profile_temp"
